Fix for #1 (when apartment share is not a integer number)

+++ UpdateFruboTenants/Owner.cs
@@ -6,7 +6,7 @@
         public string Phone { get; set; }
         public string Email { get; set; }
         public string[] WayOfInfo { get; set; }
-        public int Share { get; set; }
+        public double Share { get; set; }
         public string MovedIn { get; set; }
         public bool RegisteredAtAddress { get; set; }
     }
